backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-celery-results==2.5.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
newsapi-python==0.2.7
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
